feat: 프리뷰용 더미 데이터 관리 객체 추가
- 프리뷰에서 `DummyData`를 사용하도록 수정했습니다.
